[AllModules]: dump commit
A commit containing all previous changes for work from school (on Friday).

+++ aurora/src/main/java/com/livajusic/marko/aurora/db_repos/GifCategoryRepo.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface GifCategoryRepo extends JpaRepository<GifCategory, Integer> {
+public interface GifCategoryRepo extends JpaRepository<GifCategory, Long> {
     Optional<GifCategory> findByCategory(String category);
 
 }
