+++ aurora/src/main/java/com/livajusic/marko/aurora/db_repos/GifRepo.java
@@ -3,5 +3,5 @@ package com.livajusic.marko.aurora.db_repos;
 import com.livajusic.marko.aurora.tables.AuroraGIF;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GifRepo extends JpaRepository<AuroraGIF, Integer> {
+public interface GifRepo extends JpaRepository<AuroraGIF, Long> {
 }
